api/requirement.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-JWT-Extended==4.0.2
flask-talisman==0.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycparser==2.20
PyJWT==2.0.1
six==1.15.0
Werkzeug==1.0.1
